Remove .zshrc
The .zshrc is edited by several applications. This makes it very
difficult to maintain the file among different systems as each system
has separate set of applications and all.

To counter this, this branch has added a init.sh which takes care of all
the custom zshrc things that are needed. The init.sh is then loaded once
during the start of the actual .zshrc (The line to load init.sh must be
added manually by the user).